Integracion del nombre comercial en odoo

+++ app/Http/Controllers/ApiOdooController.php
@@ -100,10 +100,12 @@ class ApiOdooController extends Controller
                                     'user_id' => $dataClient['user_id'],
                                     'client_odoo_id' => $dataClient['id'],
                                 ]);
-                                foreach ($dataClient['tradename'] as $value) {
-                                    $client->tradenames()->create([
-                                        'name' => $value['customer_label']
-                                    ]);
+                                if (count($dataClient['tradename']) > 0) {
+                                    foreach ($dataClient['tradename'] as $value) {
+                                        $client->tradenames()->create([
+                                            'name' => $value['customer_label']
+                                        ]);
+                                    }
                                 }
                             } else {
                                 $client->update([
@@ -113,11 +115,13 @@ class ApiOdooController extends Controller
                                     'contact' => $dataClient['contact'] == false ? "Sin Contacto" : $dataClient['contact'],
                                     'user_id' => $dataClient['user_id'],
                                 ]);
-                                $client->tradenames()->delete();
-                                foreach ($dataClient['tradename'] as $value) {
-                                    $client->tradenames()->create([
-                                        'name' => $value['customer_label']
-                                    ]);
+                                if (count($dataClient['tradename']) > 0) {
+                                    $client->tradenames()->delete();
+                                    foreach ($dataClient['tradename'] as $value) {
+                                        $client->tradenames()->create([
+                                            'name' => $value['customer_label']
+                                        ]);
+                                    }
                                 }
                             }
                         }
